cannot delete worker if have registration

+++ src/Controller/Api/Formation/TeamController.php
@@ -156,6 +156,10 @@ class TeamController extends AbstractController
     {
         $em = $this->doctrine->getManager();
 
+        if(count($obj->getRegistrations()) != 0){
+            return $apiResponse->apiJsonResponseBadRequest("Cette personne est ou a été inscrit à une formation. Vous ne pouvez pas le supprimer.");
+        }
+
         $em->remove($obj);
         $em->flush();
